requirements.txt
python-telegram-bot==11.1.0
requests==2.24.0
enum34==1.1.10